# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sea Gull Restaurant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: List products
      parameters:
        - name: categoryId
          in: query
          schema:
            type: number
        - name: search
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: bestSeller
          in: query
          schema:
            type: boolean
        - name: vegetarian
          in: query
          schema:
            type: boolean
        - name: healthy
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: Get user orders
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/cancel:
    post:
      operationId: cancelOrder
      tags: [orders]
      summary: Cancel order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    RegisterRequest:
      type: object
      required: [name, phone]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        password:
          type: string
    LoginRequest:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, phone, role, loyaltyPoints]
      properties:
        id:
          type: number
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        role:
          type: string
        avatarUrl:
          type: string
        loyaltyPoints:
          type: number
        preferredLanguage:
          type: string
        darkMode:
          type: boolean
...
    Product:
      type: object
      required: [id, name, price, categoryId]
      properties:
        id:
          type: number
        categoryId:
          type: number
        name:
          type: string
        nameAr:
          type: string
        description:
          type: string
        price:
          type: string
        discountedPrice:
          type: string
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        prepTime:
          type: number
        spiceLevel:
          type: string
        isAvailable:
          type: boolean
        isFeatured:
          type: boolean
        isBestSeller:
          type: boolean
        isNew:
          type: boolean
        isVegetarian:
          type: boolean
        isHealthy:
          type: boolean
        tags:
          type: array
          items:
            type: string
        rating:
          type: string
        reviewCount:
          type: number
        calories:
          type: number
    ProductDetail:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          properties:
            ingredients:
              type: string
            allergens:
              type: string
            addOns:
              type: array
              items:
                $ref: "#/components/schemas/AddOn"
            variants:
              type: array
              items:
                $ref: "#/components/schemas/ProductVariant"
            relatedProducts:
              type: array
              items:
                $ref: "#/components/schemas/Product"
    AddOn:
      type: object
      required: [id, name, price]
      properties:
        id:
          type: number
        name:
          type: string
        nameAr:
          type: string
        price:
          type: string
        isAvailable:
          type: boolean
    ProductVariant:
      type: object
      required: [id, name]
      properties:
        id:
          type: number
        name:
          type: string
        nameAr:
          type: string
        priceDiff:
          type: string
        isDefault:
          type: boolean
    Order:
      type: object
      required: [id, orderNumber, status, total, createdAt]
      properties:
        id:
          type: number
        orderNumber:
          type: string
        status:
          type: string
        type:
          type: string
        subtotal:
          type: string
        deliveryFee:
          type: string
        serviceFee:
          type: string
        tax:
          type: string
        tip:
          type: string
        discount:
          type: string
        total:
          type: string
        deliveryAddress:
          type: string
        deliveryNotes:
          type: string
        specialInstructions:
          type: string
        couponCode:
          type: string
        loyaltyPointsEarned:
          type: number
        loyaltyPointsUsed:
          type: number
        estimatedDeliveryTime:
          type: string
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        statusHistory:
          type: array
          items:
            $ref: "#/components/schemas/StatusHistory"
        branch:
          $ref: "#/components/schemas/Branch"
        rider:
          $ref: "#/components/schemas/RiderInfo"
    OrderItem:
      type: object
      required: [id, productName, quantity, unitPrice, totalPrice]
      properties:
        id:
          type: number
        productId:
          type: number
        productName:
          type: string
        productImage:
          type: string
        variantName:
          type: string
        quantity:
          type: number
        unitPrice:
          type: string
        totalPrice:
          type: string
        addOns:
          type: array
          items:
            type: object
        specialInstructions:
          type: string
    StatusHistory:
      type: object
      required: [status, createdAt]
      properties:
        status:
          type: string
        note:
          type: string
        createdAt:
          type: string
    RiderInfo:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        phone:
          type: string
        rating:
          type: string
        vehicleType:
          type: string
        profileImageUrl:
          type: string
        currentLatitude:
          type: string
        currentLongitude:
          type: string
    CreateOrderRequest:
      type: object
      required: [branchId, items, paymentMethod, type, subtotal, total]
      properties:
        branchId:
          type: number
        items:
          type: array
          items:
            type: object
        paymentMethod:
          type: string
        type:
          type: string
        addressId:
          type: number
        deliveryAddress:
          type: string
        deliveryNotes:
          type: string
        specialInstructions:
          type: string
        couponCode:
          type: string
        loyaltyPointsToUse:
          type: number
        tip:
          type: number
        subtotal:
          type: number
        deliveryFee:
          type: number
        serviceFee:
          type: number
        tax:
          type: number
        total:
          type: number
        scheduledFor:
          type: string
        isContactless:
          type: boolean
...
    Branch:
      type: object
      required: [id, name, address]
      properties:
        id:
          type: number
        name:
          type: string
        nameAr:
          type: string
        address:
          type: string
        area:
          type: string
        city:
          type: string
        phone:
          type: string
        latitude:
          type: string
        longitude:
          type: string
        isActive:
          type: boolean
        acceptsDelivery:
          type: boolean
        acceptsPickup:
          type: boolean
...